requirements.txt
customtkinter==5.2.2
pygame==2.6.1
Pillow==10.3.0
opencv-python==4.13.0.90
mediapipe==0.10.32
numpy==2.2.6
